+++ test/integration/sequelize.test.js
@@ -252,6 +252,7 @@ describe(Support.getTestDialectTeaser('Sequelize'), function() {
       return this.sequelize.query(this.insertQuery);
     });
 
+
     describe('logging', function () {
       it('executes a query with global benchmarking option and default logger', function() {
         var logger = sinon.spy(console, 'log');
@@ -263,6 +264,24 @@ describe(Support.getTestDialectTeaser('Sequelize'), function() {
         return sequelize.query('select 1;').then(function() {
           expect(logger.calledOnce).to.be.true;
           expect(logger.args[0][0]).to.be.match(/Executed \(default\): select 1; Elapsed time: \d+ms/);
+        });
+      });
+    
+      it('logs warnings when there are warnings', function() {
+        var logger = sinon.spy();
+        var sequelize = Support.createSequelizeInstance({
+          logging: logger,
+          benchmark: false,
+          showWarnings: true
+        });
+        var insertWarningQuery = 'INSERT INTO ' + qq(this.User.tableName) + ' (username, email_address, ' +
+          qq('createdAt') + ', ' + qq('updatedAt') +
+          ") VALUES ('john', 'john@gmail.com', 'HORSE', '2012-01-01 10:10:10')";
+
+        return sequelize.query(insertWarningQuery)
+        .then(function(results) {
+          expect(logger.callCount).to.equal(3);
+          expect(logger.args[2][0]).to.be.match(/^MySQL Warnings \(default\): Data truncated for column 'createdAt'/m);
         });
       });